join checkers modules
