+++ CCAgT_utils/formats/mask.py
@@ -9,7 +9,7 @@ from PIL import Image
 
 from CCAgT_utils.base.categories import CategoriesInfos
 from CCAgT_utils.base.errors import ShapeError
-from CCAgT_utils.formats.checkers import is_2d
+from CCAgT_utils.checkers import is_2d
 
 
 @dataclass
